ml-models/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
albumentations>=1.3.0
tensorboard>=2.13.0
wandb>=0.15.0
transformers>=4.30.0
timm>=0.9.0 